function/requirements.txt
google-cloud-error-reporting==0.30.0
requests
